+++ conftest.py
@@ -85,6 +85,19 @@ def practitioner():
 
 
 @pytest.fixture
+def related_person(patient):
+    related_person = koppeltaal.models.RelatedPerson(
+        patient=patient,
+        relationship='PRN',
+        name=koppeltaal.models.Name(
+                family=[u"Doe"],
+                given=[u"John"],
+                text=u'John Doe'),
+        gender="M")
+    return related_person
+
+
+@pytest.fixture
 def activity_definition(connector):
     # Highly depending on what's activated on the server.
     definition = connector.activity('KTSTESTGAME')